Create test_animation.cpp
